+++ controller/server/view/common/icon.mli
@@ -7,3 +7,4 @@ val screen : [> Html_types.svg ] Tyxml.Html.elt
 val document : [> Html_types.svg ] Tyxml.Html.elt
 val arrow_left : [> Html_types.svg ] Tyxml.Html.elt
 val letter : [> Html_types.svg ] Tyxml.Html.elt
+val copyright : [> Html_types.svg ] Tyxml.Html.elt